feat: added service implementation
